Fixed the resorting issue.
Added a way to define your own sorting function. Since there is no way
to sort the array of objects without changing the ordering, a custom
sorting function must be supplied if the order is important. This
custom sorting function MUST sort by the parent value first, or it
won’t display properly.